tests/requirements-dev.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --extra dev --output-file=tests/requirements-dev.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   openai
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via typer
coverage==7.6.10
    # via time-guardian (pyproject.toml)
distro==1.9.0
    # via openai
exceptiongroup==1.2.2
    # via
    #   anyio
    #   pytest
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
jiter==0.8.2
    # via openai
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
mss==10.0.0
    # via time-guardian (pyproject.toml)
mypy==1.14.1
    # via time-guardian (pyproject.toml)
mypy-extensions==1.0.0
    # via mypy
numpy==2.2.1
    # via time-guardian (pyproject.toml)
openai==1.59.7
    # via time-guardian (pyproject.toml)
packaging==24.2
    # via
    #   pytest
    #   pytest-sugar
pillow==11.1.0
    # via time-guardian (pyproject.toml)
pluggy==1.5.0
    # via pytest
pydantic==2.10.5
    # via openai
pydantic-core==2.27.2
    # via pydantic
pygments==2.19.1
    # via rich
pytest==8.3.4
    # via
    #   time-guardian (pyproject.toml)
    #   pytest-randomly
    #   pytest-sugar
pytest-randomly==3.16.0
    # via time-guardian (pyproject.toml)
pytest-sugar==1.0.0
    # via time-guardian (pyproject.toml)
requests==2.32.3
    # via time-guardian (pyproject.toml)
rich==13.9.4
    # via typer
ruff==0.9.2
    # via time-guardian (pyproject.toml)
schedule==1.2.2
    # via time-guardian (pyproject.toml)
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   openai
termcolor==2.5.0
    # via pytest-sugar
tomli==2.2.1
    # via
    #   mypy
    #   pytest
tqdm==4.67.1
    # via openai
typer==0.15.1
    # via time-guardian (pyproject.toml)
typing-extensions==4.12.2
    # via
    #   anyio
    #   mypy
    #   openai
    #   pydantic
    #   pydantic-core
    #   rich
    #   typer
urllib3==2.3.0
    # via requests
wheel==0.45.1
    # via time-guardian (pyproject.toml)
